Rename to AB Testing, since it's more an A/B testing library than a true multivariate library.
