+++ .idea/modules.xml
@@ -3,11 +3,11 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/Multivariate.iml" filepath="$PROJECT_DIR$/Multivariate.iml" />
+      <module fileurl="file://$PROJECT_DIR$/ab-annotations/multivariate-annotations.iml" filepath="$PROJECT_DIR$/ab-annotations/multivariate-annotations.iml" />
+      <module fileurl="file://$PROJECT_DIR$/ab-compiler/multivariate-compiler.iml" filepath="$PROJECT_DIR$/ab-compiler/multivariate-compiler.iml" />
+      <module fileurl="file://$PROJECT_DIR$/ab-lib/ab-lib.iml" filepath="$PROJECT_DIR$/ab-lib/ab-lib.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/multivariate-annotations/multivariate-annotations.iml" filepath="$PROJECT_DIR$/multivariate-annotations/multivariate-annotations.iml" />
       <module fileurl="file://$PROJECT_DIR$/multivariate-annotations/multivariate-api.iml" filepath="$PROJECT_DIR$/multivariate-annotations/multivariate-api.iml" />
-      <module fileurl="file://$PROJECT_DIR$/multivariate-compiler/multivariate-compiler.iml" filepath="$PROJECT_DIR$/multivariate-compiler/multivariate-compiler.iml" />
-      <module fileurl="file://$PROJECT_DIR$/multivariate-lib/multivariate-lib.iml" filepath="$PROJECT_DIR$/multivariate-lib/multivariate-lib.iml" />
     </modules>
   </component>
 </project>
