+++ ab-annotations/src/main/java/com/afewroosloose/abtesting/api/DefinedTest.java
@@ -1,0 +1,8 @@
+package com.afewroosloose.abtesting.api;
+
+/**
+ * This interface is used for creating tests to pass into the MultivariateProcessor.
+ */
+public interface DefinedTest {
+  void run();
+}
