+++ ab-annotations/src/main/java/com/afewroosloose/abtesting/api/annotations/ResourceTest.java
@@ -1,0 +1,17 @@
+package com.afewroosloose.abtesting.api.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@TesterAnnotation
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ResourceTest {
+  String testName();
+
+  String method() default "setText";
+
+  int[] values() default {};
+}
