+++ ab-annotations/src/main/java/com/afewroosloose/abtesting/api/annotations/TesterAnnotation.java
@@ -1,0 +1,14 @@
+package com.afewroosloose.abtesting.api.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Created by matt on 22/05/2016.
+ */
+@Target(ElementType.ANNOTATION_TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface TesterAnnotation {
+}
