+++ ab-compiler/build.gradle
@@ -1,0 +1,53 @@
+apply plugin: 'java'
+apply plugin: 'maven'
+
+group='com.github.imperial-crystalline-recursion'
+
+sourceCompatibility = 1.7
+targetCompatibility = 1.7
+
+buildscript {
+  repositories {
+    jcenter()
+  }
+}
+repositories {
+  jcenter()
+}
+
+dependencies {
+  compile fileTree(dir: 'libs', include: ['*.jar'])
+  compile 'com.google.auto.service:auto-service:1.0-rc2'
+  compile project(path: ':ab-annotations')
+  compile 'com.squareup:javapoet:1.7.0'
+  testCompile 'junit:junit:4.12'
+  testCompile 'com.google.testing.compile:compile-testing:0.9'
+  testCompile files(org.gradle.internal.jvm.Jvm.current().getToolsJar()) //gradle kinda sucks so we need this
+}
+
+task sourcesJar(type: Jar, dependsOn: classes) {
+  classifier = 'sources'
+  from sourceSets.main.allSource
+}
+
+artifacts {
+  archives sourcesJar
+}
+
+Task task = task writePom {
+  pom {
+    project {
+      inceptionYear '2016'
+      version '0.2'
+      licenses {
+        license {
+          name 'The Apache Software License, Version 2.0'
+          url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
+          distribution 'repo'
+        }
+      }
+    }
+  }.writeTo("$buildDir/pom/pom.xml")
+}
+
+task.mustRunAfter(assemble)
