+++ ab-lib/src/main/AndroidManifest.xml
@@ -1,0 +1,13 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.afewroosloose.abtesting.lib"
+    >
+
+  <application
+      android:allowBackup="true"
+      android:label="@string/app_name"
+      android:supportsRtl="true"
+      >
+
+  </application>
+
+</manifest>
