+++ ab-lib/src/main/java/com/afewroosloose/abtesting/lib/TestPicker.java
@@ -1,0 +1,8 @@
+package com.afewroosloose.abtesting.lib;
+
+/**
+ * Created by matt on 2/07/2016.
+ */
+public abstract class TestPicker {
+  public abstract int choose(int numberOfTests);
+}
