+++ ab-lib/src/main/res/values/strings.xml
