+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
   buildToolsVersion "23.0.3"
 
   defaultConfig {
-    applicationId "com.afewroosloose.multivariate"
+    applicationId "com.afewroosloose.abtesting"
     minSdkVersion 8
     targetSdkVersion 23
     versionCode 1
@@ -34,7 +34,7 @@ dependencies {
   testCompile 'junit:junit:4.12'
   testCompile "org.robolectric:robolectric:3.0"
   compile 'com.android.support:appcompat-v7:23.3.0'
-  apt project(':multivariate-compiler')
-  compile project(path: ':multivariate-annotations')
-  compile project(path: ':multivariate-lib')
+  apt project(':ab-compiler')
+  compile project(path: ':ab-annotations')
+  compile project(path: ':ab-lib')
 }
