+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.afewroosloose.multivariate"
+    package="com.afewroosloose.abtesting"
     >
 
   <application
@@ -10,7 +10,7 @@
       android:supportsRtl="true"
       android:theme="@style/AppTheme"
       >
-    <activity android:name=".MainActivity">
+    <activity android:name="com.afewroosloose.abtesting.MainActivity">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
