--- multivariate-annotations/build.gradle
@@ -1,38 +1,0 @@
-apply plugin: 'java'
-apply plugin: 'com.github.dcendents.android-maven'
-
-group='com.github.imperial-crystalline-recursion'
-
-sourceCompatibility = 1.7
-targetCompatibility = 1.7
-
-dependencies {
-  compile fileTree(dir: 'libs', include: ['*.jar'])
-}
-
-task sourcesJar(type: Jar, dependsOn: classes) {
-  classifier = 'sources'
-  from sourceSets.main.allSource
-}
-
-artifacts {
-  archives sourcesJar
-}
-
-Task task = task writePom {
-  pom {
-    project {
-      inceptionYear '2016'
-      version '0.1'
-      licenses {
-        license {
-          name 'The Apache Software License, Version 2.0'
-          url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-          distribution 'repo'
-        }
-      }
-    }
-  }.writeTo("$buildDir/pom/pom.xml")
-}
-
-task.mustRunAfter(assemble)
