--- multivariate-annotations/src/main/java/com/afewroosloose/multivariate/api/AbstractTest.java
@@ -1,10 +1,0 @@
-package com.afewroosloose.multivariate.api;
-
-/**
- * This is for internal use. We use this to generate tests using the annotation processor.
- */
-public abstract class AbstractTest {
-  public abstract void run(int run);
-
-  public abstract int getNumberOfTests();
-}
