--- multivariate-annotations/src/main/java/com/afewroosloose/multivariate/api/DefinedTest.java
@@ -1,8 +1,0 @@
-package com.afewroosloose.multivariate.api;
-
-/**
- * This interface is used for creating tests to pass into the MultivariateProcessor.
- */
-public interface DefinedTest {
-  void run();
-}
