--- multivariate-annotations/src/main/java/com/afewroosloose/multivariate/api/annotations/ResourceTest.java
@@ -1,17 +1,0 @@
-package com.afewroosloose.multivariate.api.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@TesterAnnotation
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ResourceTest {
-  String testName();
-
-  String method() default "setText";
-
-  int[] values() default {};
-}
