--- multivariate-lib/build.gradle
@@ -1,57 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'com.github.dcendents.android-maven'
-
-group='com.github.imperial-crystalline-recursion'
-
-android {
-  compileSdkVersion 23
-  buildToolsVersion "23.0.3"
-
-  defaultConfig {
-    minSdkVersion 8
-    targetSdkVersion 23
-    versionCode 1
-    versionName "1.0"
-  }
-  buildTypes {
-    release {
-      minifyEnabled false
-      proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-    }
-  }
-}
-
-dependencies {
-  compile fileTree(dir: 'libs', include: ['*.jar'])
-  testCompile 'junit:junit:4.12'
-  compile project(path: ':multivariate-annotations')
-}
-
-// build a jar with source files
-task sourcesJar(type: Jar) {
-  from android.sourceSets.main.java.srcDirs
-  classifier = 'sources'
-}
-
-artifacts {
-  archives sourcesJar
-}
-
-Task task = task writePom {
-  pom {
-    project {
-      inceptionYear '2016'
-      version '0.1'
-      licenses {
-        license {
-          name 'The Apache Software License, Version 2.0'
-          url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-          distribution 'repo'
-        }
-      }
-    }
-  }.writeTo("$buildDir/pom/pom.xml")
-}
-
-task.mustRunAfter(assemble)
-
