--- multivariate-lib/src/main/java/com/afewroosloose/multivariate/lib/SplitPicker.java
@@ -1,11 +1,0 @@
-package com.afewroosloose.multivariate.lib;
-
-/**
- * Picks a test using a uniform random distribution
- */
-public class SplitPicker extends TestPicker {
-  @Override
-  public int choose(int numberOfTests) {
-    return (int) (Math.random() * Integer.MAX_VALUE) % numberOfTests;
-  }
-}
