--- multivariate-lib/src/main/java/com/afewroosloose/multivariate/lib/TestPicker.java
@@ -1,8 +1,0 @@
-package com.afewroosloose.multivariate.lib;
-
-/**
- * Created by matt on 2/07/2016.
- */
-public abstract class TestPicker {
-  public abstract int choose(int numberOfTests);
-}
