--- multivariate-lib/src/test/java/com/afewroosloose/multivariate/lib/ExampleUnitTest.java
@@ -1,15 +1,0 @@
-package com.afewroosloose.multivariate.lib;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * To work on unit tests, switch the Test Artifact in the Build Variants view.
- */
-public class ExampleUnitTest {
-  @Test
-  public void addition_isCorrect() throws Exception {
-    assertEquals(4, 2 + 2);
-  }
-}
